Extract juno endow-related withdraws to separate hook
